# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/users/me:
    get:
      tags:
      - user-rest-controller
      summary: Shows the logged user's profile
      operationId: getUser
      parameters:
      - name: pageItineraries
        in: query
        description: itineraries page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageReviews
        in: query
        description: reviews page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "400":
          description: Operation not available when no user is logged in
        "200":
          description: Displayed profile properly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
      - user-rest-controller
      summary: Edits the user's information
      operationId: editUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "400":
          description: Email does not match against correct mail format
        "403":
          description: No user is logged in
        "200":
          description: Successfully edited user's data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "500":
          description: Empty body of request
...
  /api/management/users/{id}:
    get:
      tags:
      - user-management-rest-controller
      summary: Get the specified user
      operationId: getUser_1
      parameters:
      - name: id
        in: path
        description: id of user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: Not authorized
        "404":
          description: User not found
    put:
      tags:
      - user-management-rest-controller
      summary: Edit existing user
      operationId: editUser_1
      parameters:
      - name: id
        in: path
        description: id of user to be edited
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: User was correctly edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: Not authorized
        "404":
          description: User not found
    delete:
      tags:
      - user-management-rest-controller
      summary: Delete user
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: id of user to be deleted
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: User was correctly deleted
        "403":
          description: Not authorized
        "404":
          description: User not found
...
  /api/itineraries/{id}/image:
    get:
      tags:
      - itinerary-rest-controller
      summary: Returns the image of the desired Itinerary.
      operationId: downloadImage_2
      parameters:
      - name: id
        in: path
        description: itinerary id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Requested a non-existing Itinerary's image.
        "200":
          description: Sucessfully returned the Itinerary's image.
          content:
            image/jpeg: {}
    put:
      tags:
      - itinerary-rest-controller
      summary: Edit the image of an itinerary
      operationId: editItineraryImage
      parameters:
      - name: id
        in: path
        description: id of the itinerary you want to add an image to
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
          image/jpeg:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
          image/png:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
      responses:
        "403":
          description: Login needed to perform operation or User is not owner of the
            itinerary
        "404":
          description: Itinerary not found
        "200":
          description: Image edited successfully
          content:
            image/jpeg: {}
...
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      summary: Returns the profile image of the desired user
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        description: user id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully returned the user image
          content:
            image/jpeg: {}
        "404":
          description: Requested a non-existing User's image.
...
  /api/places/{id}/image:
    get:
      tags:
      - place-rest-controller
      summary: Returns the image of the desired Place.
      operationId: downloadImage_1
      parameters:
      - name: id
        in: path
        description: place id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Requested a non-existing Place's image.
        "200":
          description: Sucessfully returned the Place's image.
          content:
            image/jpeg: {}
...
  /api/destinations/{id}/image:
    get:
      tags:
      - destination-rest-controller
      summary: Returns the image of the desired Destination.
      operationId: downloadImage_3
      parameters:
      - name: id
        in: path
        description: destination id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully returned the Destination's image.
          content:
            image/jpeg: {}
        "404":
          description: Requested a non-existing Destination's image.
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        nationality:
          type: string
        image:
          type: boolean
        roles:
          type: array
          items:
            type: string
    UserDTO:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        passwordHash:
          type: string
        nationality:
          type: string
      description: new user's information
...